Implemented basic printing logic.
The Print() method is clearly not workable. It blocks the main thread, and the printer will not respond until the command is executed, which may be several minutes.

Command-sending and response-awaiting must be moved to either a seperate thread or a timer which periodically polls for responses and sends new commands if appropriate.